+++ modules/misc.js
@@ -54,6 +54,12 @@ export class MaestroConfigForm extends FormApplication {
      * Provide data to the template
      */
     getData() {
+        const criticalSuccessFailureTracks = game.settings.get(MAESTRO.MODULE_NAME, MAESTRO.SETTINGS_KEYS.Misc.criticalSuccessFailureTracks);
+
+        if (!this.data && criticalSuccessFailureTracks) {
+            this.data = criticalSuccessFailureTracks;
+        }
+        
         return {
             playlists: game.playlists.entities,
             criticalSuccessPlaylist: this.data.criticalSuccessPlaylist,
@@ -108,7 +114,7 @@ export class MaestroConfigForm extends FormApplication {
 function _addPlaylistLoopToggle(html) {
     const playlistModeButtons = html.find('[data-action="playlist-mode"]');
     const loopToggleHtml = 
-        `<a class="sound-control" data-action="playlist-loop" title="${game.i18n.localize("PLAYLIST-LOOP.ButtonTooltipLoop")}">
+        `<a class="sound-control" data-action="playlist-loop" title="${game.i18n.localize("MAESTRO.PLAYLIST-LOOP.ButtonTooltipLoop")}">
             <i class="fas fa-sync"></i>
         </a>`;
 
@@ -136,10 +142,10 @@ function _addPlaylistLoopToggle(html) {
         const mode = playlist.data.mode;
         if ([-1, 2].includes(mode)) {
             button.setAttribute("class", buttonClass.concat(" disabled"));
-            button.setAttribute("title", game.i18n.localize("PLAYLIST-LOOP.ButtonToolTipDisabled"));
+            button.setAttribute("title", game.i18n.localize("MAESTRO.PLAYLIST-LOOP.ButtonToolTipDisabled"));
         } else if (loop === false) {
             button.setAttribute("class", buttonClass.concat(" inactive"));
-            button.setAttribute("title", game.i18n.localize("PLAYLIST-LOOP.ButtonTooltipNoLoop"));
+            button.setAttribute("title", game.i18n.localize("MAESTRO.PLAYLIST-LOOP.ButtonTooltipNoLoop"));
         }
     }
 
@@ -161,11 +167,11 @@ function _addPlaylistLoopToggle(html) {
         if (buttonClass.includes("inactive")) {
             game.playlists.get(playlistId).unsetFlag(MAESTRO.MODULE_NAME, MAESTRO.DEFAULT_CONFIG.PlaylistLoop.flagNames.loop);
             button.setAttribute("class", buttonClass.replace(" inactive", ""));
-            button.setAttribute("title", game.i18n.localize("PLAYLIST-LOOP.ButtonTooltipLoop"));
+            button.setAttribute("title", game.i18n.localize("MAESTRO.PLAYLIST-LOOP.ButtonTooltipLoop"));
         } else { 
             game.playlists.get(playlistId).setFlag(MAESTRO.MODULE_NAME, MAESTRO.DEFAULT_CONFIG.PlaylistLoop.flagNames.loop, false);
             button.setAttribute("class", buttonClass.concat(" inactive"));
-            button.setAttribute("title", game.i18n.localize("PLAYLIST-LOOP.ButtonTooltipNoLoop"));
+            button.setAttribute("title", game.i18n.localize("MAESTRO.PLAYLIST-LOOP.ButtonTooltipNoLoop"));
         }
     });
 }
@@ -176,38 +182,42 @@ function _addPlaylistLoopToggle(html) {
  * @param {*} update 
  * @todo maybe return early if no flag set?
  */
-export function _onPreUpdatePlaylistSound(playlist, update) {
-    // Return if there's no id or the playlist is not in sequential or shuffl mode
-    if (!playlist.data.playing || !update._id || ![0, 1].includes(playlist.data.mode)) {
-        return;
+export function _onPreUpdatePlaylistSound(sound, update, options, userId) {
+    // skip this method if the playlist sound has already been processed
+    if (sound?._maestroSkip) return true;
+
+    sound._maestroSkip = true;
+    const playlist = sound.parent;
+    // Return if there's no id or the playlist is not in sequential or shuffle mode
+    if (!playlist?.data?.playing || !update?._id || ![0, 1].includes(playlist?.data?.mode)) {
+        return true;
     }
 
     // If the update is a sound playback ending, save it as the previous track and return
-    if (update.playing === false) {
-        return playlist.setFlag(MAESTRO.MODULE_NAME, MAESTRO.DEFAULT_CONFIG.PlaylistLoop.flagNames.previousSound, update._id);
+    if (update?.playing === false) {
+        playlist.setFlag(MAESTRO.MODULE_NAME, MAESTRO.DEFAULT_CONFIG.PlaylistLoop.flagNames.previousSound, update._id);
+        return true;
     }
 
     // Otherwise it must be a sound playback starting:
     const previousSound = playlist.getFlag(MAESTRO.MODULE_NAME, MAESTRO.DEFAULT_CONFIG.PlaylistLoop.flagNames.previousSound);
 
-    if (!previousSound) {
-        return;
-    }
+    if (!previousSound) return true;
 
     let order;
 
     // If shuffle order exists, use that, else map the sounds to an order
-    if (playlist.data.mode === 1) {
-        order = playlist._getPlaybackOrder();
+    if (playlist?.data?.mode === 1) {
+        order = playlist.playbackOrder;
     } else {
-        order = playlist.sounds.map(s => s._id);
+        order = playlist?.sounds.map(s => s._id);
     }        
     
     const previousIdx = order.indexOf(previousSound);
     const playlistloop = playlist.getFlag(MAESTRO.MODULE_NAME, MAESTRO.DEFAULT_CONFIG.PlaylistLoop.flagNames.loop);
 
     // If the previous sound was the last in the order, and playlist loop is set to false, don't play the incoming sound
-    if (previousIdx === (playlist.sounds.length - 1) && playlistloop === false) {
+    if (previousIdx === (playlist?.sounds?.length - 1) && playlistloop === false) {
         update.playing = false;
         playlist.data.playing = false;
     }        
@@ -219,7 +229,7 @@ export function _onPreUpdatePlaylistSound(playlist, update) {
 export function _onPreCreateChatMessage(message, options, userId) {
     const removeDiceSound = game.settings.get(MAESTRO.MODULE_NAME, MAESTRO.SETTINGS_KEYS.Misc.disableDiceSound);
 
-    if (removeDiceSound && (message.sound && message.sound === "sounds/dice.wav")) {
+    if (removeDiceSound && message.sound === "sounds/dice.wav") {
         message.sound = "";
     }
 }
@@ -244,15 +254,17 @@ export function _onRenderChatMessage(message, html, data) {
  * @param {*} message
  */
 function playCriticalSuccessFailure(message) {
-    if ( !game.user.isGM || !message.isRoll || !message.isRollVisible || !message.roll.parts.length ) return;
+    if ( !game.user.isGM || !message.isRoll || !message.isContentVisible || !message.isRoll ) return;
   
-    // Evaluate rolls where the first part is a d20 roll
-    let d = message.roll.parts[0];
-    const isD20Roll = d instanceof Die && (d.faces === 20) && (d.results.length === 1);
-    if ( !isD20Roll ) return;
-  
-    // Ensure it is not a modified roll
-    const isModifiedRoll = ("success" in d.rolls[0]) || d.options.marginSuccess || d.options.marginFailure;
+    // Highlight rolls where the first part is a d20 roll
+    const roll = message.roll;
+    if ( !roll.dice.length ) return;
+    const d = roll.dice[0];
+
+    // Ensure it is an un-modified d20 roll
+    const isD20 = (d.faces === 20) && ( d.results.length === 1 );
+    if ( !isD20 ) return;
+    const isModifiedRoll = ("success" in d.results[0]) || d.options.marginSuccess || d.options.marginFailure;
     if ( isModifiedRoll ) return;
 
     // Get the sounds
@@ -263,9 +275,9 @@ function playCriticalSuccessFailure(message) {
     const criticalFailureSound = criticalSuccessFailureTracks.criticalFailureSound;
        
     // Play relevant sound for successes and failures
-    if ((d.total >= (d.options.critical || 20)) && (criticalSuccessPlaylist && criticalSuccessSound)) {
+    if ((d.options.critical && (d.total >= d.options.critical)) && (criticalSuccessPlaylist && criticalSuccessSound)) {
         Playback.playTrack(criticalSuccessSound, criticalSuccessPlaylist);
-    } else if ((d.total <= (d.options.fumble || 1)) && (criticalFailurePlaylist && criticalFailureSound)) {
+    } else if ((d.options.fumble && (d.total <= d.options.fumble)) && (criticalFailurePlaylist && criticalFailureSound)) {
         Playback.playTrack(criticalFailureSound, criticalFailurePlaylist)
     }
 }