+++ CHANGELOG.md
@@ -1,15 +1,59 @@
 # Changelog
 
-## [Unreleased]
-- Soundboard for SFX
-- Possible support for external audio services (YouTube etc)
+## Known Issues
+1. Critical Sounds do not play for game systems that do not use the same roll logic as dnd5e
+2. Critical Sounds do not play when roll made with advantage/disadvantage
+3. Critical Sounds do not play with modules such as BetterRolls5e
 
-## [0.6.5] - 2020-04-26
-## Changed
+
+## [0.7.5] - 2021-07-13
+- Added Foundry VTT v0.8.x compatibility
+
+## [0.7.4] - 2021-02-01
+- Release workflow refinements
+
+## [0.7.3] - 2021-01-31
+- Skip processing playlist sound updates if already processed
+- Improve some logic in Hype Track
+- Improve some logic in Item Track
+
+## [0.7.2] - 2020-10-31
+### Added
+- Português (Brasil) translation (thanks @rinnocenti!)
+- Spanish translation (thanks @lozalojo)
+
+### Fixed
+- Combat Track config not opening
+
+## [0.7.1] - 2020-06-20
+### Added
+- 한국어 (Korean) translation (thanks KLO!)
+
+### Fixed
+- Fixed bug where Combat Tracks playlist was not created when first enabled
+- Maestro additional config now opens from the Module Settings menu
+
+## [0.7.0] - 2020-06-19
+### Added
+- New macro methods:
+- - `game.maestro.findSound` finds a playlist sound by its name or path
+- - `game.maestro.playByName` allows playing a playlist sound using its name (best effort)
+- - `game.maestro.pause` pauses one or many sounds (or names of sounds)
+- - `game.maestro.pauseAll` pauses all active playlist sounds
+- - `game.maestro.resume` resumes playing one or many sounds
+- - `game.maestro.playHype` play an actor's hype track
+- New setting to allow Hype Tracks to pause other playing sounds (disabled by default) -- paused sounds will resume when the Hype Track finishes
+
+### Changed
+- Players can now set **Hype Tracks** for Actors they own
 - Refactored hooks to use new structure
+- Certify compatibility with Foundry VTT 0.6.3
+- Maestro Config is now also available in Module Settings
 
-## Fixed
+### Fixed
 - Fixed bug where Hype Tracks could not be created/set when first enabled
+- Fixed bug where Hype Tracks playlist was not created when first enabled
+- Fixed bug where Item Tracks playlist was not created when first enabled
 
 ## [0.6.4] - 2020-04-18
 ### Fixed